+++ spring/src/main/java/ru/yoomoney/tech/dbqueue/spring/dao/PostgresQueueDao.java
@@ -53,7 +53,7 @@ public class PostgresQueueDao implements QueueDao {
         MapSqlParameterSource params = new MapSqlParameterSource()
                 .addValue("queueName", location.getQueueId().asString())
                 .addValue("payload", enqueueParams.getPayload())
-                .addValue("executionDelay", enqueueParams.getExecutionDelay().getSeconds());
+                .addValue("executionDelay", enqueueParams.getExecutionDelay().toMillis());
 
         queueTableSchema.getExtFields().forEach(paramName -> params.addValue(paramName, null));
         enqueueParams.getExtData().forEach(params::addValue);
@@ -81,7 +81,7 @@ public class PostgresQueueDao implements QueueDao {
                 new MapSqlParameterSource()
                         .addValue("id", taskId)
                         .addValue("queueName", location.getQueueId().asString())
-                        .addValue("executionDelay", executionDelay.getSeconds()));
+                        .addValue("executionDelay", executionDelay.toMillis()));
         return updatedRows != 0;
     }
 
@@ -97,7 +97,7 @@ public class PostgresQueueDao implements QueueDao {
                         queueTableSchema.getExtFields().stream().collect(Collectors.joining(", ", ", ", ""))) +
                 ") VALUES " +
                 "(" + location.getIdSequence().map(seq -> "nextval('" + seq + "'), ").orElse("") +
-                ":queueName, :payload, now() + :executionDelay * INTERVAL '1 SECOND', 0, 0" +
+                ":queueName, :payload, now() + :executionDelay * INTERVAL '1 MILLISECOND', 0, 0" +
                 (queueTableSchema.getExtFields().isEmpty() ? "" : queueTableSchema.getExtFields().stream()
                         .map(field -> ":" + field).collect(Collectors.joining(", ", ", ", ""))) +
                 ") RETURNING " + queueTableSchema.getIdField();
@@ -110,7 +110,7 @@ public class PostgresQueueDao implements QueueDao {
 
     private String createReenqueueSql(@Nonnull QueueLocation location) {
         return "UPDATE " + location.getTableName() + " SET " + queueTableSchema.getNextProcessAtField() +
-                " = now() + :executionDelay * INTERVAL '1 SECOND', " +
+                " = now() + :executionDelay * INTERVAL '1 MILLISECOND', " +
                 queueTableSchema.getAttemptField() + " = 0, " +
                 queueTableSchema.getReenqueueAttemptField() +
                 " = " + queueTableSchema.getReenqueueAttemptField() + " + 1 " +